Fixed height problem + space-around
The height problem happened becouse flex apply the rules to every direct
son, and the direct sons of the div.pollsList are the <poll> elements,
not the div.poll elements (which are INSIDE the <poll> elements).
The solution is to apply the container style (border, shadow and
bgcolor) to the <poll> element instead of the div.poll element.
I don't really like to style a custom html element like <poll> but I
can't find any other solution.